fix(refactoring): move tests into seperate dir
