+++ .eslintignore
@@ -1,1 +1,1 @@
-src/__mocks__
+tests/__mocks__
