+++ package.json
@@ -26,8 +26,8 @@
     "prepublishOnly": "run-p build",
     "start": "node dist/index.js",
     "test": "run-s test:lint test:types test:unit test:self",
-    "test:lint": "eslint ./src  --ext js,ts,tsx",
-    "test:lint:fix": "eslint ./src  --ext js,ts,tsx --fix",
+    "test:lint": "eslint ./tests  --ext js,ts,tsx",
+    "test:lint:fix": "eslint ./src ./tests --ext js,ts,tsx --fix",
     "test:self": "scriptlint",
     "test:self:fix": "scriptlint --fix",
     "test:types": "tsc -p tsconfig.json --noEmit true",
