+++ tests/__mocks__/cosmiconfig.js
