+++ tests/__mocks__/fs.js
