+++ tsconfig.json
@@ -66,5 +66,5 @@
 		/* Advanced Options */
 		"forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
 	},
-	"include": ["./src"]
+	"include": ["./src", "./tests"]
 }
